refactor(persist): use pick and omit from stapp-core
