+++ packages/stapp/src/helpers/omit/omit.spec.ts
@@ -1,6 +1,6 @@
 import { omit } from './omit'
 
-describe('utils/object/omit', () => {
+describe('omit', () => {
   it('should omit props', () => {
     expect(omit({ a: 1, b: 2 }, ['a'])).toEqual({ b: 2 })
     expect(omit({ a: 2 }, ['a', 'c', 'd'] as any)).toEqual({})
